Add patch StopWavingThatFlashlight.
When entering a conversation while holding a flashlight, it's supposed to stop the IK animation of the flashlight, but an "on idle anim stop" callback is executed very frequently for some reason. Work around it by assuming that if the "light" is still enabled, we probably aren't in the idle anim (which turns off the light), and we don't need to run the "on idle anim stop" callback.

Now, the flashlight will go to a default position. (Ideally, it would "freeze" where it was.)

Currently not working for the (old) GOG patch.

The actual code path which is supposed to disable IK:
- ConversationLoggerPageSystem::StopWorldOnConversationStart()
- Character.ToggleFlashlightIK()
- FlashlightBehaviour.stopIK()